jasman-backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.95.2
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
frozenlist==1.6.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
gotrue==2.12.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mtcnn==1.0.0
multidict==6.4.3
namex==0.0.9
networkx==3.4.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==5.29.4
pyasn1==0.6.1
pydantic==1.10.21
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
realtime==2.4.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.27.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf_keras==2.19.0
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.22.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
SQLAlchemy==2.0.0
apscheduler